chore(aider): Enable gitignore by default
This commit changes the default value of the `gitignore` setting in the `.aider.conf.yml`
file to `true`.

This change ensures that `.aider*` files are added to `.gitignore` by default, preventing them
from being accidentally committed to the repository. This is a common practice to keep the
repository clean and avoid tracking unnecessary files.