refactor(aliases): Refactor aichat build functions
This commit refactors the aichat build function into three separate functions. The new functions
are `_aichat_setup_functions`, `_aichat_build_agents` and `_aichat_build_tools`.

The reason for this change is to make the code more modular and easier to understand. It also
allows for more flexibility in the future.