fix(.aliases): Use ipv4 alias to get city for wttr
This commit fixes an issue where the `wttr` alias was not working as expected due to the `ip`
alias returning an unexpected value. The `ip` alias has been replaced with the `ipv4` alias,
which returns the correct value for the city.

The `ipv4` alias is more reliable than the `ip` alias, as it specifically returns the IPv4
address of the machine. This ensures that the `wttr` alias will always return the weather for
the correct location.