chore(config): Update package lock file
This commit updates the versions of awscli and uv in the config.lock file.

The awscli version is updated from 2.23.11 to 2.23.12, and the uv version is updated from 0.5.26
to 0.5.27. The checksums for these new versions are also updated. This ensures that the correct
versions of these tools are installed and that their integrity is verified.