chore(config): Update tool versions in mise config lock
This commit updates the versions of several tools in the `mise` configuration lock file. The
updated tools include `awscli`, `slsa-verifier`.

The reason for these updates is to ensure that the project is using the latest versions of
these tools, which may include bug fixes, performance improvements, or new features. Updating
the lock file ensures that these versions are consistent across different environments.  </chore>