chore(config): Update cargo-binstall and cargo-update
This commit updates the versions of `cargo-binstall` and `cargo-update` in the
`config.lock` file.

`cargo-binstall` is updated from version 1.12.1 to 1.12.2, and `cargo-update` is
updated from version 16.2.1 to 16.3.0. The checksum for the `cargo-binstall` binary
has also been updated to reflect the new version. These updates ensure that the latest
versions of these tools are used.