fix(dotfiles): enable loading of .env file in nixd alias
This change removes a comment to actually enable the loading of the .env file when running the nixd alias. This allows environment variables defined in .env files to be respected.

The change was made to fix a previously disabled feature that was intended to read .env files and pass them as environment variables to the container.